requirements.txt
requests==2.18.4
Django==2.1.2
hashids==1.2.0
beautifulsoup4==4.7.0
djangorestframework==3.9.0
